requirements.txt

streamlit
sentence-transformers
faiss-cpu
wikipedia-api
tqdm
